app: stack frames support
Use information collected by the parser about subprogram calls to
properly track BPF state across different stacks.

When a state is computed for each log line, and a a jump instruction
of kind BPF2BPF_CALL is encountered, a copy of the base (previous)
state is saved, and a new clean state is created (to emulate fresh
stack frame). Scratch registers are copied, as this is how function
arguments are passed into the subprogram.

When an exit instruction is encountered, the last saved state is
restored, scratching callee saved registers and copying r0 from the
previous state.

Track location information in BpfState and display it in the state
panel: current frame, idx and pc of the instruction associated with
the state.

Implement indentation of the instruction stream based on current stack
frame. Emphasize call/exit instructions a start/end of a block.

Remove throttling of handleScroll, as it appears to be causing visual
issues.

Fix keyboard scroll when selected line moves out of sight.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>