app: switch to full log render
Render the entire log and rely on automatic scroll bar for navigation.

So far only the visible part of the log was added to the DOM for
browser to render. This was achieved by tracking topLineIdx - the
index of the first line to render - and then relying on it as an
anchor across the codebase. The reason for this design was to ensure
that huge logs (100+ Mb) are loaded and rendered smoothly.

A significant UI/UX downside of this approach is the necessity to
manually handle scrolling. Initial dumb implementation simply handled
a delta based on user input events. It worked fine, but this way there
is no actual scrollbar in the UI, which is confusing.

Later I implemented a virtual scroll bar by adding a "scroll-tape"
element with huge height, and then positioning log lines on that tape
to emulate the scrolling. This required explicit handling of scroll
events, which turned out to be complicated and led to all sorts of
problems both in terms of the implementation and the UI.

With this change, the entire log is processed and div elements are
added for each input line on load. This required significant changes
in how updateView() and keyboard navigation works:
  * caching divs is not necessary: they are all in the DOM
  * topLineIdx is no longer a rendering anchor, logContainer.scroll is
  * line-numbers columns are now rendered just once

updateView() still needs to be smart in order for the app to remain
interactive: frequent style updates are performed only for a part of
the log, which is visible (or close to it).

This new approach makes navigation UX better. The cost of it is that
the responsivness of the app will degrade with increased size of the
input log. However, as it stands, verifier log level 1 in practice is
small enough. And support of level 2 is not a priority yet.

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>