app: implement filling out of source code view
After the input log is loaded, collect recognized C_SOURCE lines into
a separate list. Then use it to populate the content of
source-code-view.

The log contains only the lines that are "executed" in the process of
verification, which means that the source code information is always
incomplete. For now populate missing lines with "nop", to make it
obvious that pieces are missing.

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>