analyzer: implement CSourceMap collection
Introduce CSourceMap class, and a field of VerifierLogState,
representing the relationships between recognized C source code lines
and BPF instructions in the log.

It contains deduplicated C source code lines and maps to and from
indexes of the original log.

CSourceLine objects are identified by "filename:lineNum" string.

Additionally, collect fileRange map, mapping a C filename to a range
of line numbers that were found in the input log.

This is enough data to be able to display all available C source
content by filename, and visualize the realtionships between
instructions and C source code.

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>