Determine number of args for calls based on known reg values
We are printing call instructions in the following format:

    r0 = call foo(r1, r2, r3, r4, r5)

This is done to enable interactivity with argument registers.

However, in most cases only a few of 5 registers are actually
used. Given that we know the values of said registers, we can check
them and if scratched then it's not a call argument.

To enable this, introduce prevValue field to BpfValue. It is set in
nextBpfState() for Effect.UPDATE.

In CallHtml component access to BpfState is necessary to access this
information, so pass it through from LogLinesRaw component.

With that we show, for example:

    r0 = call foo(r1, r2, r3)

In case r4 and r5 are scratched before the call.