Add C Source Line View
This adds a collapsible C source line view
on the right, which contains all C files
from the log and their C lines.
This view is in sync with the main log lines
view; meaning if a user clicks on a line
in one, we'll hightlight the associated lines
in the other.

This also removes C lines from the main
log view by default but a checkbox was added
so users can add them back in as desired.

Additional changes:
- remove line numbers from the main log view
- remove Go To Line input box
- Add C Line to State Panel