Handle "global and assumed valid" messages
Verifier may emit a message indicating a global function call, which
is not verified inline, for example:

    565: (85) call pc+234
    Func#5 ('pick_any_bit') is global and assumed valid.

This is important for visualizer, because so far `call pc+offset`
pattern was handled exclusively as subprogram call.

A small challenge with this is that the overall design of bpfvv
expects line-by-line processing of the input log. And to handle cases
like this some backtracking (or lookahead) mechanism is necessary,
even though it is convenient to work with "lines" fundamentally.

Address this with the following changes:
* introduce ParsedLineType.KNOWN_MESSAGE and KnownMessageInfo to
  recognize particular verifier messages
* leave parser interface unchanged (parseLine)
* split parsed lines processing in the analyzer into separate steps
  * parse lines individually
  * process known messages, potentially editing parsedLines
    array (this is backtracking essentially)
  * and then build bpfStates and cSourceMap