analyzer.ts: fix dependencies through side effects
Commit 2d07fba enhanced dependency tracking with information about
value changes reported by the verifier. This has broken an assumption
in `getMemSlotDependencies` that only one slot can be written by an
instruction. With side effects, it can be many. This lead to incorrect
dependency chain being displayed in some cases.

Fix this by distinguishing and additional case in
`getMemSlotDependencies`, which now handles three:
  * an actual update such as from `+=` instruction
  * an actual write, which can be detected by checking `writes` set of
    a `BpfInstruction`
  * a side effect, which can only be noticed via value change in the
    BpfState (this is the new case)