Dropdown and load spinner (#85)
* When a new log is loaded scroll to last ins

Instead of scrolling and selected the last
line which is usually junk, scroll to the
last ins which is right above the error msg.

* Make a dropdown list of example logs

Load this from a global javascript object
so we can easily change it from a server
env.

Also add a loading spinner when the
user pastes a new log, loads a file,
or loads an example log.

Issue:
- https://github.com/libbpf/bpfvv/issues/80